+++ src/feishu/CardRenderer.ts
@@ -2395,7 +2395,12 @@ function codeBlock(value: string, maxLength: number): string {
 }
 
 function stripAnsi(value: string): string {
-  return value.replace(/\u001B\[[0-?]*[ -/]*[@-~]/g, "");
+  return value
+    .replace(/\u001B\][\s\S]*?(?:\u0007|\u001B\\|$)/g, "")
+    .replace(/\u001B[P^_X][\s\S]*?(?:\u001B\\|$)/g, "")
+    .replace(/\u001B\[[0-?]*[ -/]*[@-~]/g, "")
+    .replace(/\u001B[@-_]/g, "")
+    .replace(/[\u0000-\u0008\u000B\u000C\u000E-\u001A\u001C-\u001F\u007F-\u009F]/g, "");
 }
 
 function inlineCode(value: string): string {